+++ repro/models/yuan2021/model.py
@@ -1,6 +1,5 @@
 import json
 import logging
-import os
 from typing import Dict, List, Tuple, Union
 
 from repro.common import util